ssh-agent: Correct logic
I was not noticing if some other process was on the SSH_AGENT_PID.
If there is an ssh-agent on our existing SSH_AGENT_PID, then we're all good.
Otherwise, start another one and write out the env file.
This does mean we can end up with an excess of agents, but they do no harm.