git: Enable --autosquash by default
Doesn't have any effect unless you make a commit with the message:

    squash! shasum
    fixup! Subject-line of target commit

When the commit log message begins with "squash! ..." (or "fixup! ..."), and
there is already a commit in the todo list that matches, automatically modifies
the todo list of rebase -i so that the commit marked for squashing comes right
after the commit to be modified, and changes the action to squash (or fixup).