Makefile: Only symlink if the link doesn't exist
Do it the make way, using pattern rules.
We specify the make targets, not the source files.
If the target doesn't exist, we symlink it.

The only exception to this pattern is .bashrc and .profile,
which (normally) already exist. We are grepping them to find out
whether to add the source line.